requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
coverage==7.2.5
Django==4.2.1
django-crispy-forms==2.0
django-jalali==6.0.1
django-profanity-filter==0.2.1
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jdatetime==4.1.1
model-bakery==1.11.0
mypy-extensions==1.0.0
outcome==1.2.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.1
pluggy==1.0.0
py==1.11.0
PySocks==1.7.1
pytest==6.2.5
pytest-base-url==2.0.0
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-html==3.2.0
pytest-metadata==2.0.4
pytest-selenium==4.0.0
pytest-variables==2.0.0
redis==4.5.5
requests==2.30.0
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
tenacity==6.3.1
toml==0.10.2
trio==0.22.0
trio-websocket==0.10.2
urllib3==2.0.2
wsproto==1.2.0

djangorestframework~=3.14.0